CHANGE: Added logging and fixed some bugs on asset sell plugin.

+++ DAP/plugin/digital_asset_transaction/fermat-dap-plugin-digital-asset-transaction-asset-seller-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/digital_asset_transaction/asset_seller/developer/bitdubai/version_1/structure/events/AssetSellerMonitorAgent.java
@@ -168,13 +168,20 @@ public class AssetSellerMonitorAgent extends FermatAgent {
             for (DAPMessage message : assetTransmission.getUnreadDAPMessageBySubject(DAPMessageSubject.NEGOTIATION_ANSWER)) {
                 try {
                     AssetNegotiationContentMessage content = (AssetNegotiationContentMessage) message.getMessageContent();
+                    NegotiationRecord record = dao.getNegotiationForId(content.getAssetNegotiation().getNegotiationId());
+                    if (record.getNegotiationStatus() == AssetSellStatus.NEGOTIATION_CANCELLED) {
+                        System.out.println("This transaction was already cancelled.");
+                        continue;
+                    }
                     SellingRecord sellingRecord = dao.getLastSellingRecord(content.getAssetNegotiation().getNegotiationId());
                     switch (content.getSellStatus()) {
                         case NEGOTIATION_CONFIRMED: {
+                            System.out.println("Negotiation confirmed...");
                             dao.updateSellingStatus(sellingRecord.getRecordId(), AssetSellStatus.NEGOTIATION_CONFIRMED);
                             break;
                         }
                         case NEGOTIATION_REJECTED: {
+                            System.out.println("Negotiation rejected...");
                             unlockFunds(sellingRecord.getMetadata());
                             dao.deleteSellingRecord(sellingRecord.getRecordId());
                             break;
@@ -202,10 +209,12 @@ public class AssetSellerMonitorAgent extends FermatAgent {
             for (SellingRecord record : dao.getActionRequiredSellings()) {
                 switch (record.getStatus()) {
                     case NEGOTIATION_CONFIRMED: {
+                        System.out.println("Negotiation confirmed...");
                         ActorAssetUser mySelf = actorAssetUserManager.getActorAssetUser();
                         CryptoAddress cryptoVaultAddress = cryptoVaultManager.getAddress(record.getMetadata().getNetworkType());
                         cryptoAddressBookManager.registerCryptoAddress(cryptoVaultAddress, mySelf.getActorPublicKey(), Actors.DAP_ASSET_USER, record.getBuyer().getActorPublicKey(), Actors.DAP_ASSET_USER, Platforms.DIGITAL_ASSET_PLATFORM, VaultType.CRYPTO_CURRENCY_VAULT, CryptoCurrencyVault.BITCOIN_VAULT.getCode(), "reference_wallet", ReferenceWallet.BASIC_WALLET_BITCOIN_WALLET);
                         DraftTransaction draftTransaction = assetVaultManager.createDraftTransaction(record.getMetadata().getLastTransactionHash(), record.getBuyer().getCryptoAddress());
+                        dao.updateSellerTransaction(record.getNegotiationId(), draftTransaction);
                         AssetSellContentMessage contentMessage = new AssetSellContentMessage(record.getRecordId(), draftTransaction.serialize(), AssetSellStatus.WAITING_FIRST_SIGNATURE, record.getMetadata(), record.getNegotiationId(), draftTransaction.getValue(), cryptoVaultAddress);
                         DAPMessage dapMessage = new DAPMessage(contentMessage, mySelf, record.getBuyer(), DAPMessageSubject.NEW_SELL_STARTED);
                         assetTransmission.sendMessage(dapMessage);
@@ -213,6 +222,7 @@ public class AssetSellerMonitorAgent extends FermatAgent {
                         break;
                     }
                     case PARTIALLY_SIGNED: {
+                        System.out.println("Negotiation partially signed...");
                         if (!Validate.isValidTransaction(record.getBuyerTransaction(), record.getSellerTransaction())) {
                             AssetSellContentMessage content = new AssetSellContentMessage(record.getRecordId(), null, AssetSellStatus.SIGNATURE_REJECTED, null, record.getNegotiationId(), record.getSellerTransaction().getValue(), null);
                             ActorAssetUser mySelf = actorAssetUserManager.getActorAssetUser();
@@ -230,11 +240,13 @@ public class AssetSellerMonitorAgent extends FermatAgent {
                         break;
                     }
                     case COMPLETE_SIGNATURE: {
+                        System.out.println("Complete signed...");
                         debitUserWallet(record, BalanceType.BOOK);
                         dao.updateSellingStatus(record.getRecordId(), AssetSellStatus.WALLET_DEBITED);
                         break;
                     }
                     case WALLET_DEBITED: {
+                        System.out.println("Broadcasting...");
                         bitcoinNetworkManager.broadcastTransaction(record.getBroadcastingTxHash());
                         dao.updateSellingStatus(record.getRecordId(), AssetSellStatus.BROADCASTING);
                     }
@@ -244,6 +256,7 @@ public class AssetSellerMonitorAgent extends FermatAgent {
                             case BROADCASTED: {
                                 debitUserWallet(record, BalanceType.AVAILABLE);
                                 dao.updateSellingStatus(record.getRecordId(), AssetSellStatus.SELL_FINISHED);
+                                System.out.println("Asset Sold...");
                                 break;
                             }
                             case WITH_ERROR: {
@@ -260,17 +273,23 @@ public class AssetSellerMonitorAgent extends FermatAgent {
         private void checkTimeout() throws DAPException, CantUpdateRecordException, CantLoadTableToMemoryException, CantDeleteRecordException, CantLoadWalletException {
             for (NegotiationRecord negotiation : dao.getWaitingConfirmationNegotiations()) {
                 if (negotiation.isExpired()) {
-                    dao.updateNegotiationStatus(negotiation.getNegotiation().getNegotiationId(), AssetSellStatus.NEGOTIATION_CANCELLED);
-                    for (SellingRecord record : dao.getAllSelingRecordsForNegotiation(negotiation.getNegotiation().getNegotiationId())) {
-                        unlockFunds(record.getMetadata());
-                        dao.deleteSellingRecord(record.getRecordId());
-                    }
-                    AssetNegotiationContentMessage content = new AssetNegotiationContentMessage(AssetSellStatus.NEGOTIATION_CANCELLED, negotiation.getNegotiation());
-                    ActorAssetUser user = actorAssetUserManager.getActorAssetUser();
-                    DAPMessage message = new DAPMessage(content, user, negotiation.getBuyer(), DAPMessageSubject.NEGOTIATION_CANCELLED);
-                    assetTransmission.sendMessage(message);
+                    cancellNegotiation(negotiation);
                 }
             }
+        }
+
+        private void cancellNegotiation(NegotiationRecord negotiation) throws DAPException, CantLoadTableToMemoryException, CantUpdateRecordException, CantDeleteRecordException, CantLoadWalletException {
+            System.out.println("Cancelling negotiation for asset: " + negotiation.getNegotiation().getAssetToOffer().getName());
+            dao.updateNegotiationStatus(negotiation.getNegotiation().getNegotiationId(), AssetSellStatus.NEGOTIATION_CANCELLED);
+            for (SellingRecord record : dao.getAllSelingRecordsForNegotiation(negotiation.getNegotiation().getNegotiationId())) {
+                unlockFunds(record.getMetadata());
+                System.out.println("Deleting selling record: " + record.getRecordId());
+                dao.deleteSellingRecord(record.getRecordId());
+            }
+            AssetNegotiationContentMessage content = new AssetNegotiationContentMessage(AssetSellStatus.NEGOTIATION_CANCELLED, negotiation.getNegotiation());
+            ActorAssetUser user = actorAssetUserManager.getActorAssetUser();
+            DAPMessage message = new DAPMessage(content, user, negotiation.getBuyer(), DAPMessageSubject.NEGOTIATION_CANCELLED);
+            assetTransmission.sendMessage(message);
         }
 
         private void debitUserWallet(SellingRecord record, BalanceType balance) throws CantGetAssetUserActorsException, CantLoadWalletException, CantGetCryptoTransactionException, CantGetTransactionsException, CantRegisterDebitException {
